SONAR-5338 When data is coming from the 'test_data' metric, time field can return a float

+++ sonar-server/src/main/java/org/sonar/server/test/ws/TestsShowAction.java
@@ -117,7 +117,8 @@ public class TestsShowAction implements RequestHandler {
 
         json.prop("name", cursor.getAttrValue("name"));
         json.prop("status", cursor.getAttrValue("status").toUpperCase());
-        json.prop("durationInMs", Long.parseLong(cursor.getAttrValue("time")));
+        // time can contain float value, we have to truncate it
+        json.prop("durationInMs", ((Double) Double.parseDouble(cursor.getAttrValue("time"))).longValue());
 
         SMInputCursor errorCursor = cursor.childElementCursor();
         if (errorCursor.getNext() != null) {
